split partial classes
